:white_check_mark: Refacto tests to make them more readable + fix some tests
