+++ .gitignore
@@ -26,3 +26,8 @@ docker-compose.override.yml
 # OS files
 .DS_Store
 Thumbs.db
+
+# Test artifacts (generated by integration tests)
+test_*
+existing_project
+custom_directory_path
