--- docker-compose.yml
@@ -1,92 +1,0 @@
-# Generated by Athena v0.1.0 from FASTAPI_PROJECT deployment
-# Project Version: 2.0.0
-# Generated: 2025-09-13 20:37:37 UTC
-# Features: Intelligent defaults, optimized networking, enhanced health checks
-# DO NOT EDIT MANUALLY - This file is auto-generated
-
-# Services: 3 configured with intelligent defaults
-
-services:
-  backend:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    container_name: fastapi-project-backend
-    ports:
-    - 8000:8000
-    environment:
-      DATABASE_URL: ${DATABASE_URL}
-      SECRET_KEY: ${SECRET_KEY}
-      REDIS_URL: ${REDIS_URL}
-    command: uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload
-    depends_on:
-    - database
-    - redis
-    healthcheck:
-      test:
-      - CMD-SHELL
-      - curl -f http://localhost:8000/health || exit 1
-      interval: 30s
-      timeout: 10s
-      retries: 3
-      start_period: 40s
-    restart: unless-stopped
-    deploy:
-      resources:
-        limits:
-          cpus: '1.0'
-          memory: 1024M
-      restart_policy:
-        condition: on-failure
-        delay: 5s
-        max_attempts: 3
-        window: 120s
-    networks:
-    - fastapi_network
-    pull_policy: missing
-    labels:
-      athena.generated: 2025-09-13
-      athena.type: generic
-      athena.project: FASTAPI_PROJECT
-      athena.service: backend
-  database:
-    image: postgres:15
-    container_name: fastapi-project-database
-    ports:
-    - 5432:5432
-    environment:
-      POSTGRES_USER: ${POSTGRES_USER}
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
-      POSTGRES_DB: ${POSTGRES_DB}
-    volumes:
-    - ./postgres-data:/var/lib/postgresql/data
-    restart: always
-    networks:
-    - fastapi_network
-    pull_policy: missing
-    labels:
-      athena.generated: 2025-09-13
-      athena.type: database
-      athena.project: FASTAPI_PROJECT
-      athena.service: database
-  redis:
-    image: redis:7-alpine
-    container_name: fastapi-project-redis
-    ports:
-    - 6379:6379
-    command: redis-server --appendonly yes
-    volumes:
-    - ./redis-data:/data:rw
-    restart: always
-    networks:
-    - fastapi_network
-    pull_policy: missing
-    labels:
-      athena.project: FASTAPI_PROJECT
-      athena.generated: 2025-09-13
-      athena.type: cache
-      athena.service: redis
-networks:
-  fastapi_network:
-    driver: bridge
-name: FASTAPI_PROJECT
