+++ src/athena/generator/defaults.rs
@@ -55,7 +55,7 @@ pub struct EnhancedDockerService {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub ports: Option<Vec<String>>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub environment: Option<HashMap<String, String>>,
+    pub environment: Option<Vec<String>>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub command: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -280,29 +280,30 @@ impl DefaultsEngine {
         Some(port_strings)
     }
     
-    fn convert_environment(env_vars: &[EnvironmentVariable]) -> Option<HashMap<String, String>> {
+    fn convert_environment(env_vars: &[EnvironmentVariable]) -> Option<Vec<String>> {
         if env_vars.is_empty() {
             return None;
         }
         
-        let mut env_map = HashMap::new();
+        let mut env_list = Vec::new();
         for env_var in env_vars {
             match env_var {
                 EnvironmentVariable::Template(var_name) => {
-                    env_map.insert(var_name.clone(), format!("${{{}}}", var_name));
+                    env_list.push(format!("{}=${{{}}}", var_name, var_name));
                 }
                 EnvironmentVariable::Literal(value) => {
-                    // Try to parse KEY=VALUE format, fallback to generic name
-                    if let Some((key, val)) = value.split_once('=') {
-                        env_map.insert(key.to_string(), val.to_string());
+                    // If it's already in KEY=VALUE format, use as-is
+                    // Otherwise, treat as a standalone value
+                    if value.contains('=') {
+                        env_list.push(value.clone());
                     } else {
-                        env_map.insert("ENV_VALUE".to_string(), value.clone());
+                        env_list.push(format!("VALUE={}", value));
                     }
                 }
             }
         }
         
-        Some(env_map)
+        Some(env_list)
     }
     
     fn convert_volumes(volumes: &[VolumeMapping]) -> Option<Vec<String>> {
