+++ tests/integration/mod.rs
@@ -3,4 +3,4 @@ pub mod cli_commands_test;
 pub mod docker_compose_generation_test;
 pub mod error_handling_test;
 pub mod boilerplate_generation_test;
-pub mod structural_tests;
+pub mod structural;
